+++ CHANGELOG.md
@@ -1,6 +1,34 @@
 Changelog
 =========
 
+## 3.12.0 (2018-01-24)
+
+### Enhancements
+
+* Deprecated the deploy API in favor of the build API
+  Calls to `deploy` will now redirect to `build`.
+  [#454](https://github.com/bugsnag/bugsnag-php/pull/454)
+
+### Fixes
+
+* Re-added apiKey parameter to payload for older API versions
+  [#459](https://github.com/bugsnag/bugsnag-php/pull/459)
+
+## 3.11.0 (2018-01-09)
+
+### Enhancements
+
+* Add support for tracking sessions and overall crash rate by setting
+  `autoCaptureSessions` in configuration options. In addition, sessions can be
+  indicated manually using `Client->startSession()`
+
+## 3.10.0 (2017-12-14)
+
+### Enhancements
+
+* Add 'addMetaData' method to reports
+  [#430](https://github.com/bugsnag/bugsnag-php/pull/430)
+
 ## 3.9.0 (2017-11-23)
 
 ### Enhancements